перенес код по загрузке изображений в отдельный модуль
